# apps/web/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wedding Website API",
    "version": "1.0.0",
    "description": "API for the wedding website - manage guests, events, RSVPs, gifts, photos, and more"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/admin/photos": {
      "get": {
        "operationId": "get-admin-photos",
        "summary": "List all photos",
        "description": "Fetch all photos including inactive ones (admin only)",
        "tags": ["Admin - Photos"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-admin-photos",
        "summary": "Create a photo",
        "description": "Add a new photo to the gallery (admin only)",
        "tags": ["Admin - Photos"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePhotoBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePhotoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/seating-charts": {
      "get": {
        "operationId": "get-admin-seating-charts",
        "summary": "List seating charts",
        "description": "Fetch all seating charts ordered by last updated",
        "tags": ["Admin - Seating"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-admin-seating-charts",
        "summary": "Create seating chart",
        "description": "Create a new seating chart with a name, default seats per table, and optional notes",
        "tags": ["Admin - Seating"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSeatingChartBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/seating-charts/{id}/generate": {
      "post": {
        "operationId": "post-admin-seating-charts-{id}-generate",
        "summary": "Generate AI seating suggestions",
        "description": "Use OpenAI to generate intelligent seating assignments based on guest relationships, sides, and family groups",
        "tags": ["Admin - Seating"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Resource UUID"
            },
            "required": true,
            "description": "Resource UUID",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSeatingBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/seating-charts/{id}/tables": {
      "post": {
        "operationId": "post-admin-seating-charts-{id}-tables",
        "summary": "Add table to seating chart",
        "description": "Create a new table in a seating chart with optional number, name, capacity, position, and shape",
        "tags": ["Admin - Seating"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Resource UUID"
            },
            "required": true,
            "description": "Resource UUID",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTableBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-admin-seating-charts-{id}-tables",
        "summary": "Delete all tables from chart",
        "description": "Bulk delete all tables from a seating chart",
        "tags": ["Admin - Seating"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Resource UUID"
            },
            "required": true,
            "description": "Resource UUID",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation succeeded"
          }
        },
        "required": ["success"]
      },
...
      "PhotoListResponse": {
        "type": "object",
        "properties": {
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhotoResponse"
            },
            "description": "List of photos"
          }
        },
        "required": ["photos"]
      },
      "CreatePhotoBody": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Photo URL (required)"
          },
          "alt": {
            "type": "string",
            "description": "Alt text (required)"
          },
          "description": {
            "type": "string",
            "description": "Photo description"
          }
        },
        "required": ["url", "alt"]
      },
      "CreatePhotoResponse": {
        "type": "object",
        "properties": {
          "photo": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PhotoResponse"
              }
            ],
            "description": "Created photo"
          }
        },
        "required": ["photo"]
      },
...
      "GenerateSeatingBody": {},
...
      "CreateTableBody": {},
      "CreateSeatingChartBody": {},
...
      "PhotoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Photo UUID"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Photo URL"
          },
          "alt": {
            "type": "string",
            "description": "Alt text"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Photo description"
          },
          "display_order": {
            "type": "number",
            "description": "Display order"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the photo is active"
          },
          "created_at": {
            "type": "string",
            "description": "Creation timestamp"
          }
        },
        "required": [
          "id",
          "url",
          "alt",
          "description",
          "display_order",
          "is_active",
          "created_at"
        ]
      },
...
    },
...
  }
}